+++ rethinkdb-net/Interfaces/IConnection.cs
@@ -26,11 +26,9 @@ namespace RethinkDb
 
         #region Asynchronous API; synchronous API is provided by extension methods
 
-        Task<T> RunAsync<T>(IDatumConverterFactory datumConverterFactory, ISingleObjectQuery<T> queryObject);
+        Task<T> RunAsync<T>(IDatumConverterFactory datumConverterFactory, IScalarQuery<T> queryObject);
 
         IAsyncEnumerator<T> RunAsync<T>(IDatumConverterFactory datumConverterFactory, ISequenceQuery<T> queryObject);
-
-        Task<TResponseType> RunAsync<TResponseType>(IDatumConverterFactory datumConverterFactory, IWriteQuery<TResponseType> queryObject);
 
         #endregion
     }